Spin up dummy Laravel application for test use
This isn't quite canonical for Laravel packages, but we're supplying
a trait that is required to be used in an Eloquent model - we can't
really work around all of the facade references, and the code
works out simpler if we supply dummy facade references to test against.

Also alter TestModel::metadata to not trip a stack overflow.