Hook up polymorphic-relation repair to relation generation
This has become necessary due to earlier patches in this sequence removing
the erroneous assumption that _foreign key_ names had to match.  That fix
captured _all_ cases of a given polymorphic relation, and promptly blew up
an invariant in Odata metadata that requires a given relation to point to
consistent types.

To do that, we've had to slide an abstract polymorphic placeholder type
under those relation ends so we don't trip that invariant.

For moment, we are explicitly not supporting multi-generational
polymorphic relation chains, as that blew up base relation generation.
Parent -> child works fine, but grandparent -> parent -> child goes boom.