Fix PrimaryKey substitution in LaravelReadQuery
When applying a non-trivial orderBy object that contained PrimaryKey
field, getResourceSet was ignoring previously-retrieved copy of target
model's primary key and trying to dig it out anew.  This doesn't work
too well on query builders, as getKeyName doesn't exist.

The simple and obvious fix is to simply cut the orderBy handling over
to use the previously-retrieved copy of keyName, and let it worry about
whether we're hanging off a model, relation, etc.