Skip non-tabled models when booting under php artisan
Previously, models that had empty fields (because of an empty or
non-existent DB table) would cause application boot to blow up unconditionally.

This included attempts to run php artisan migrate and thus _fix_ the lack of
table - resulting in a deadlock that had to be resolved by manually editing
the model in question, migrating, and then reversing the edit.

Now, such empty models are simply skipped when running under artisan, allowing
migrate commands to be run and fix lack-of-table problem.