Hook metadata mask up to getter processing.
As Laravel allows you to exclude getters from serialisation, our metadata
processing needs to also do the same.  Previously, it was paying attention
for the actual table fields, but then blithely ignoring restrictions
for any defined getters - this can be problematic if, for example, you have
added a deprecation warning and app boot chokes on that.