Improve integration and fix request (#168)
I have tried using this library and there are a couple of things that could be improved, in particular the following:

Allow eloquent models to define metadata explicitly (I am working with mongodb, so the current implementation of inferring fields from column names is not appropiate, since mongo does not have a rigid structure).
There is a helper method in laravel's Application class to detect the app namespace, so using the PODATA_LARAVEL_APP_ROOT_NAMESPACE constant is not necessary. Other than this, Laravel's convention is to use config files and env variables instead of php constants, so that could also be improved in the future.
Other than those improvements above, which you are free to accept or not, there are a couple of fixes as well that I think are critical because they break the integration. Those two are:

Requiring symfony/yaml dependency (which is not required by laravel at the moment, so adding this package to an empty Laravel app breaks due to a missing dependency).
Setting status code in ODataController class. The current implementation is calling the setStatusCode method incorrectly with a string, breaking the code at runtime since that method only accepts integers. In fact that same status code is already being passed correctly on the previous line when calling the constructor, so that line is entirely unecessary. Maybe in previous versions of Laravel this was working, but on the current version of Laravel 5.6 that code breaks.

* Add symfony/yaml dependency to composer.json

* Fix response status code in ODataController

* Add laravel namespace helper

* Allow eloquent models to define odata metadata explicitly

* Use lower version of symfony/yaml to support php 5

* Use lower version of symfony/yaml to be compatible with other symfony packages

* Adapt MetadataProvider namespace detection to work in tests

* Add test for explicit models metadata generation

* Add model for explicit metadata definition tests