Type-guard model-specific casts.
Although cast types are only _defined_ for strings, there's nothing
stopping J. Random plonking in something else, such as an array, which
would previously cause the boot process to blow up.

Following the whitelist-over-blacklist approch in the rest of this PR,
specific casts are now only attempted if and only if the cast type is a string.